xbackend/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
astroid==2.9.0
atomicwrites==1.4.0
attrs==21.2.0
cachetools==5.0.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
colorama==0.4.4
coverage==5.5
cycler==0.10.0
Django==3.2.6
django-cors-headers==3.5.0
django-environ==0.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
future==0.18.2
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-secret-manager==2.8.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
html5lib==1.1
identify==2.4.0
idna==3.2
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
libcst==0.4.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.2
packaging==21.0
Pillow==8.3.1
platformdirs==2.4.0
pluggy==0.13.1
proto-plus==1.19.9
protobuf==3.19.4
psycopg2==2.9.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-pythonpath==0.7.3
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
reportlab==3.6.5
requests==2.26.0
rsa==4.8
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sqlparse==0.4.1
threadpoolctl==3.0.0
toml==0.10.2
typing-inspect==0.7.1
typing_extensions==4.0.1
urllib3==1.26.6
webencodings==0.5.1
wrapt==1.13.3
xhtml2pdf==0.2.5
arabic-reshaper==2.1.3
asgiref==3.4.1
astroid==2.9.0
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
colorama==0.4.4
coverage==5.5
cycler==0.10.0
Django==3.2.6
django-bootstrap4==21.2
django-cors-headers==3.5.0
django-environ==0.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
environ==1.0
future==0.18.2
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-secret-manager==2.8.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
html5lib==1.1
identify==2.4.0
idna==3.2
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
libcst==0.4.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.2
packaging==21.0
Pillow==8.3.1
platformdirs==2.4.0
pluggy==0.13.1
proto-plus==1.19.9
protobuf==3.19.4
psycopg2==2.9.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-pythonpath==0.7.3
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
reportlab==3.6.5
requests==2.26.0
rsa==4.8
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.1
threadpoolctl==3.0.0
toml==0.10.2
typing-extensions==4.0.1
typing-inspect==0.7.1
urllib3==1.26.6
webencodings==0.5.1
wrapt==1.13.3
xhtml2pdf==0.2.5